+++ dart/constraint/dynamic_joint_constraint.cpp
@@ -80,7 +80,7 @@ void DynamicJointConstraint::setErrorAllowance(double allowance)
     DART_WARN(
         "Error reduction parameter[{}] is lower than 0.0. It is set to 0.0.",
         allowance);
-    mErrorAllowance = 0.0;
+    allowance = 0.0;
   }
 
   mErrorAllowance = allowance;
@@ -100,13 +100,13 @@ void DynamicJointConstraint::setErrorReductionParameter(double erp)
     DART_WARN(
         "Error reduction parameter [{}] is lower than 0.0. It is set to 0.0.",
         erp);
-    mErrorReductionParameter = 0.0;
+    erp = 0.0;
   }
   if (erp > 1.0) {
     DART_WARN(
         "Error reduction parameter [{}] is greater than 1.0. It is set to 1.0.",
         erp);
-    mErrorReductionParameter = 1.0;
+    erp = 1.0;
   }
 
   mErrorReductionParameter = erp;
@@ -127,7 +127,7 @@ void DynamicJointConstraint::setMaxErrorReductionVelocity(double erv)
         "Maximum error reduction velocity [{}] is lower than 0.0. It is set to "
         "0.0.",
         erv);
-    mMaxErrorReductionVelocity = 0.0;
+    erv = 0.0;
   }
 
   mMaxErrorReductionVelocity = erv;
@@ -148,7 +148,7 @@ void DynamicJointConstraint::setConstraintForceMixing(double cfm)
         "Constraint force mixing parameter [{}] is lower than 1e-9. It is set "
         "to 1e-9.",
         cfm);
-    mConstraintForceMixing = 1e-9;
+    cfm = 1e-9;
   }
 
   mConstraintForceMixing = cfm;