+++ dart/constraint/joint_constraint.cpp
@@ -88,7 +88,7 @@ void JointConstraint::setErrorAllowance(double allowance)
     DART_WARN(
         "Error reduction parameter [{}] is lower than 0.0. It is set to 0.0.",
         allowance);
-    mErrorAllowance = 0.0;
+    allowance = 0.0;
   }
 
   mErrorAllowance = allowance;
@@ -108,13 +108,13 @@ void JointConstraint::setErrorReductionParameter(double erp)
     DART_WARN(
         "Error reduction parameter [{}] is lower than 0.0. It is set to 0.0.",
         erp);
-    mErrorReductionParameter = 0.0;
+    erp = 0.0;
   }
   if (erp > 1.0) {
     DART_WARN(
         "Error reduction parameter [{}] is greater than 1.0. It is set to 1.0.",
         erp);
-    mErrorReductionParameter = 1.0;
+    erp = 1.0;
   }
 
   mErrorReductionParameter = erp;
@@ -135,7 +135,7 @@ void JointConstraint::setMaxErrorReductionVelocity(double erv)
         "Maximum error reduction velocity [{}] is lower than 0.0. It is set to "
         "0.0.",
         erv);
-    mMaxErrorReductionVelocity = 0.0;
+    erv = 0.0;
   }
 
   mMaxErrorReductionVelocity = erv;
@@ -156,7 +156,7 @@ void JointConstraint::setConstraintForceMixing(double cfm)
         "Constraint force mixing parameter [{}] is lower than 1e-9. It is set "
         "to 1e-9.",
         cfm);
-    mConstraintForceMixing = 1e-9;
+    cfm = 1e-9;
   }
 
   mConstraintForceMixing = cfm;